Couple changes:
Added a ALU_ADD_DEADLINE for ALU operations with deadline instruction. When this is used, it completely ignores rs1. Since I am using the rs1 bits for deadload and deadbranch, we have to ignore the op1 value. Based on the macros defined in deadline.h, it seems we don't use it anyway (it's always g0), so i just completely removed adding op1. Whenever ALU_ADD_DEADLINE is selected as the ALU op, it just passed the value of op2 to the results.

Added the macros DEADBRANCH and DEADLOADBRANCH in soft/deadline.h

only SET the dt_status if it's a deadbranch, or else we don't care. Now the originally hacky function _check_deadlines because usable and not a hack

git-svn-id: svn+ssh://source.eecs.berkeley.edu/home/svn/chess/pret/trunk/pret_iss@1603 2fe1bebc-546c-4154-8a90-0a651ce2c4e0